Fixed a couple of crashes when using incorrect input in LanguageCourseQueries.java

+++ src/main/java/org/example/LanguageCourseQueries.java
@@ -1,5 +1,7 @@
 package org.example;
 
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.NoResultException;
 import org.example.entities.LanguageCourse;
 import org.example.entities.School;
 
@@ -27,6 +29,8 @@ public class LanguageCourseQueries {
     private static void nrOfExamsInACourse() {
         System.out.println("Enter ID of course");
         int id = getId();
+        if (id == -1)
+            return;
         inTransaction(entityManager -> {
             LanguageCourse course = entityManager.find(LanguageCourse.class,id);
             System.out.println("There are "+course.getExaminations().size()+" examinations in the "+course.getCourseName()+" course");
@@ -36,8 +40,14 @@ public class LanguageCourseQueries {
     private static void nrOfStudentsInACourse() {
         System.out.println("Enter ID of course");
         int id = getId();
+        if (id == -1)
+            return;
         inTransaction(entityManager -> {
             LanguageCourse course = entityManager.find(LanguageCourse.class,id);
+            if (course == null){
+                System.out.println("Incorrect id, returning to menu");
+                return;
+            }
             System.out.println("There are "+course.getStudents().size()+" students in the "+course.getCourseName()+" course");
         });
     }
@@ -57,6 +67,10 @@ public class LanguageCourseQueries {
         int id = getId();
         inTransaction(entityManager -> {
             LanguageCourse course = entityManager.find(LanguageCourse.class,id);
+            if (course == null) {
+                System.out.println("Incorrect course ID, returning to menu");
+                return;
+            }
             entityManager.remove(course);
         });
     }
@@ -64,20 +78,24 @@ public class LanguageCourseQueries {
     public static void updateCourseQueries() {
         System.out.println("Enter ID of course you want to update");
         int id = getId();
+        if (id == -1)
+            return;
         System.out.println("Enter new name for course:");
         String name = scanner.nextLine();
         System.out.println("Enter new start date for course:");
         String startDate = scanner.nextLine();
         System.out.println("Enter new end date for course:");
         String endDate = scanner.nextLine();
-        System.out.println("Enter new school ID for course:");
+        System.out.println("Enter new school ID for course(0 to skip):");
         int schoolId = getId();
 
-
         inTransaction(entityManager -> {
-            LanguageCourse course = entityManager.find(LanguageCourse.class,id);
             School school = entityManager.find(School.class,schoolId);
-
+            LanguageCourse course = entityManager.find(LanguageCourse.class,id);
+            if (course == null) {
+                System.out.println("Incorrect course ID, returning to menu");
+                return;
+            }
             course.setCourseName(name);
             course.setCourseStartDate(startDate);
             course.setCourseEndDate(endDate);
@@ -97,7 +115,8 @@ public class LanguageCourseQueries {
         String endDate = scanner.nextLine();
         System.out.println("Enter school ID");
         int id = getId();
-
+        if (id == -1)
+            return;
         inTransaction(entityManager -> {
             School school = entityManager.find(School.class,id);
             LanguageCourse course = new LanguageCourse();
@@ -112,8 +131,9 @@ public class LanguageCourseQueries {
 
     private static int getId() {
         int id = getChoice();
-        if (id == 100)
-            id = 0;
+        if (id == -1) {
+            System.out.println("Returning to menu");
+        }
         return id;
     }
 
@@ -140,12 +160,8 @@ public class LanguageCourseQueries {
         String name = scanner.nextLine();
 
         inTransaction(entityManager -> {
-            String queryString = """
-                    select c from LanguageCourse c where c.courseName like :name
-                    """;
-            var query = entityManager.createQuery(queryString, LanguageCourse.class);
-            query.setParameter("name", name);
-            LanguageCourse course = query.getSingleResult();
+            LanguageCourse course = getCourseByName(entityManager, name);
+            if (course == null) return;
             course.getExaminations().forEach(System.out::println);
         });
     }
@@ -155,12 +171,8 @@ public class LanguageCourseQueries {
         String name = scanner.nextLine();
 
         inTransaction(entityManager -> {
-            String queryString = """
-                    select c from LanguageCourse c where c.courseName like :name
-                    """;
-            var query = entityManager.createQuery(queryString, LanguageCourse.class);
-            query.setParameter("name", name);
-            LanguageCourse course = query.getSingleResult();
+            LanguageCourse course = getCourseByName(entityManager, name);
+            if (course == null) return;
             course.getStudents().forEach(System.out::println);
         });
     }
@@ -170,14 +182,26 @@ public class LanguageCourseQueries {
         String name = scanner.nextLine();
 
         inTransaction(entityManager -> {
-            String queryString = """
-                    select c from LanguageCourse c where c.courseName like :name
-                    """;
-            var query = entityManager.createQuery(queryString, LanguageCourse.class);
-            query.setParameter("name", name);
-            LanguageCourse course = query.getSingleResult();
+            LanguageCourse course = getCourseByName(entityManager, name);
+            if (course == null) return;
             System.out.println(course);
         });
+    }
+
+    private static LanguageCourse getCourseByName(EntityManager entityManager, String name) {
+        String queryString = """
+                select c from LanguageCourse c where c.courseName like :name
+                """;
+        var query = entityManager.createQuery(queryString, LanguageCourse.class);
+        query.setParameter("name", name);
+        LanguageCourse course;
+        try {
+            course = query.getSingleResult();
+        } catch (NoResultException e) {
+            System.out.println("Incorrect course name, returning to menu");
+            return null;
+        }
+        return course;
     }
 
     private static void showAllCoursesWithLeaderAndTeacher() {
